chore: split blls files
